+++ src/Anexia.MathematicalProgram/SolverConfiguration/SolverParameter.cs
@@ -16,15 +16,15 @@ namespace Anexia.MathematicalProgram.SolverConfiguration;
 /// <param name="TimeLimitInMilliseconds">Time limit of the solving process.</param>
 /// <param name="NumberOfThreads">The number of threads that should be used by the solver.</param>
 /// <param name="RelativeGap">The relative gap when the solver should terminate.</param>
-/// <param name="AdditionalSolverSpecificParameters">Additional solver specific parameters (key-value string, use key:value for GLOP, key=value for other supported solvers) to pass to the solver. The correct format for the desired solver
-/// must be used. Check corresponding solver documentations to be sure.</param>
+/// <param name="AdditionalSolverSpecificParameters">Additional solver specific parameters (key-value string pairs) to pass to the solver. The correct format for the desired solver
+/// must be used. Check corresponding solver documentations to be sure. E.g., Gurobi parameters can be found here: https://docs.gurobi.com/projects/optimizer/en/current/reference/parameters.html#secparameterreference</param>
 /// <param name="ExportModelFilePath">The file path and name of a file where the model should be written to.</param>
 public record SolverParameter(
     EnableSolverOutput EnableSolverOutput,
     RelativeGap? RelativeGap = null,
     TimeLimitInMilliseconds? TimeLimitInMilliseconds = null,
     NumberOfThreads? NumberOfThreads = null,
-    IReadOnlyCollection<string>? AdditionalSolverSpecificParameters = null,
+    IReadOnlyCollection<(string Key, string Value)>? AdditionalSolverSpecificParameters = null,
     string? ExportModelFilePath = null)
 {
     private const string RelativeGapKey = "RELATIVE_GAP";
@@ -109,19 +109,22 @@ public record SolverParameter(
     {
     }
 
-    internal string ToSolverSpecificParameters(IlpSolverType solverType)
+    internal List<(string Key, string Value)> ToSolverSpecificParametersList(IlpSolverType solverType)
     {
-        var parameters = new List<string>();
+        var parameters = new List<(string Key, string Value)>();
         if (NumberOfThreads is not null)
-            parameters.Add(
-                $"{IlpParameterKeyMapping[(solverType, NumberOfThreadsKey)]}={NumberOfThreads.Value}");
+            parameters.Add((IlpParameterKeyMapping[(solverType, NumberOfThreadsKey)],
+                NumberOfThreads.Value.ToString(CultureInfo.InvariantCulture)));
         if (RelativeGap is not null)
-            parameters.Add(
-                $"{IlpParameterKeyMapping[(solverType, RelativeGapKey)]}={RelativeGap.Value.ToString(CultureInfo.InvariantCulture)}");
+            parameters.Add((IlpParameterKeyMapping[(solverType, RelativeGapKey)],
+                RelativeGap.Value.ToString(CultureInfo.InvariantCulture)));
         if (AdditionalSolverSpecificParameters is not null) parameters.AddRange(AdditionalSolverSpecificParameters);
 
-        return string.Join(',', parameters);
+        return parameters;
     }
+
+    internal string ToSolverSpecificParameters(IlpSolverType solverType) => string.Join(',',
+        ToSolverSpecificParametersList(solverType).Select(parameter => $"{parameter.Key}={parameter.Value}"));
 
     internal string ToSolverSpecificParameters(LpSolverType solverType)
     {
@@ -130,7 +133,9 @@ public record SolverParameter(
             parameters.Add(
                 $"{LpParameterKeyMapping[(solverType, NumberOfThreadsKey)]}{LpKeyValueSeparators[solverType]}{NumberOfThreads.Value}");
 
-        if (AdditionalSolverSpecificParameters is not null) parameters.AddRange(AdditionalSolverSpecificParameters);
+        if (AdditionalSolverSpecificParameters is not null)
+            parameters.AddRange(AdditionalSolverSpecificParameters.Select(parameter =>
+                $"{parameter.Key}{LpKeyValueSeparators[solverType]}{parameter.Value}"));
 
         return string.Join(',', parameters);
     }